Tune Dependabot GitHub Actions grouping and ignore rule
